+++ geo/CHANGES.md
@@ -41,6 +41,8 @@
   implemented for `CoordsIter` without re-allocating (e.g., creating a `MultiPoint`).
 * Add `compose_many` method to `AffineOps`
   * <https://github.com/georust/geo/pull/1148>
+* Add `TriangulateDelaunay` and `VoronoiDiagram` traits.
+  * <https://github.com/georust/geo/pull/1144>
 
 ## 0.27.0