+++ geo/CHANGES.md
@@ -1,6 +1,9 @@
 # Changes
 
 ## Unreleased
+
+## 0.28.0
+
 * BREAKING: The `HasKernel` trait was removed and it's functionality was merged
   into `GeoNum`. If you are using common scalars for your geometry (f32, f64,
   i64, i32, i16, isize), this should have no effect on you. If you are using an